Conflicts:
	.project
	projects/OG-Analytics/.checkstyle
	projects/OG-Analytics/.classpath
	projects/OG-Analytics/.project
	projects/OG-Bloomberg/.checkstyle
	projects/OG-Bloomberg/.classpath
	projects/OG-Bloomberg/.project
	projects/OG-BloombergExample/.checkstyle
	projects/OG-BloombergExample/.classpath
	projects/OG-BloombergExample/.project
	projects/OG-BloombergExample/develop/BloombergExample server - fullstack-bloombergexample.launch
	projects/OG-Component/.checkstyle
	projects/OG-Component/.classpath
	projects/OG-Component/.project
	projects/OG-Core/.checkstyle
	projects/OG-Core/.classpath
	projects/OG-Core/.project
	projects/OG-Engine/.checkstyle
	projects/OG-Engine/.classpath
	projects/OG-Engine/.project
	projects/OG-Engine/src/main/java/com/opengamma/engine/value/ComputedValueResult.java
	projects/OG-Examples/.checkstyle
	projects/OG-Examples/.classpath
	projects/OG-Examples/.project
	projects/OG-Examples/develop/Example server - fullstack-example.launch
	projects/OG-Financial/.checkstyle
	projects/OG-Financial/.classpath
	projects/OG-Financial/.project
	projects/OG-Install/.classpath
	projects/OG-Install/.project
	projects/OG-Integration/.checkstyle
	projects/OG-Integration/.classpath
	projects/OG-Integration/.project
	projects/OG-Language/.checkstyle
	projects/OG-Language/.classpath
	projects/OG-Language/.project
	projects/OG-LiveData/.checkstyle
	projects/OG-LiveData/.classpath
	projects/OG-LiveData/.project
	projects/OG-Master/.checkstyle
	projects/OG-Master/.classpath
	projects/OG-Master/.project
	projects/OG-MasterDB/.checkstyle
	projects/OG-MasterDB/.classpath
	projects/OG-MasterDB/.project
	projects/OG-MasterDB/src/main/java/com/opengamma/batch/domain/CalculationConfiguration.hbm.xml
	projects/OG-MasterDB/src/main/java/com/opengamma/batch/domain/ComputeHost.hbm.xml
	projects/OG-MasterDB/src/main/java/com/opengamma/batch/domain/ComputeNode.hbm.xml
	projects/OG-MasterDB/src/main/java/com/opengamma/batch/domain/HbComputationTargetSpecification.hbm.xml
	projects/OG-MasterDB/src/main/java/com/opengamma/batch/domain/LiveDataField.hbm.xml
	projects/OG-MasterDB/src/main/java/com/opengamma/batch/domain/MarketData.hbm.xml
	projects/OG-MasterDB/src/main/java/com/opengamma/batch/domain/MarketDataValue.hbm.xml
	projects/OG-MasterDB/src/main/java/com/opengamma/batch/domain/RiskRunProperty.hbm.xml
	projects/OG-MasterDB/src/main/java/com/opengamma/masterdb/batch/DbBatchMaster.elsql
	projects/OG-MasterDB/src/main/java/com/opengamma/masterdb/batch/DbBatchWriter.elsql
	projects/OG-Maths/.checkstyle
	projects/OG-Maths/.classpath
	projects/OG-Maths/.project
	projects/OG-MavenPlugin/.checkstyle
	projects/OG-MavenPlugin/.classpath
	projects/OG-Parent/pom.xml
	projects/OG-Provider/.checkstyle
	projects/OG-Provider/.classpath
	projects/OG-Provider/.project
	projects/OG-Security/.checkstyle
	projects/OG-Security/.classpath
	projects/OG-Security/.project
	projects/OG-Security/src/main/java/com/opengamma/security/auditlog/AuditLogEntry.hbm.xml
	projects/OG-Security/src/main/java/com/opengamma/security/user/Authority.hbm.xml
	projects/OG-Security/src/main/java/com/opengamma/security/user/UserGroup.hbm.xml
	projects/OG-Util/.checkstyle
	projects/OG-Util/.classpath
	projects/OG-Util/.project
	projects/OG-Web/.checkstyle
	projects/OG-Web/.classpath
	projects/OG-Web/.project

+++ projects/OG-Financial/src/test/java/com/opengamma/financial/InMemoryConventionBundleMasterTest.java
@@ -44,7 +44,7 @@ public class InMemoryConventionBundleMasterTest {
     AssertJUnit.assertEquals(InMemoryConventionBundleMaster.IN_MEMORY_UNIQUE_SCHEME.getName(), uidON.getScheme());
     AssertJUnit.assertEquals(actact, conventions.getDayCount());
     AssertJUnit.assertEquals(following, conventions.getBusinessDayConvention());
-    AssertJUnit.assertEquals(0, conventions.getSettlementDays());
+    AssertJUnit.assertEquals(0, (int) conventions.getSettlementDays());
 
     final ConventionBundle conventions2 = source.getConventionBundle(ExternalId.of(InMemoryConventionBundleMaster.SIMPLE_NAME_SCHEME, "USD LIBOR 3m"));
     AssertJUnit.assertEquals("USD LIBOR 3m", conventions2.getName());
@@ -54,7 +54,7 @@ public class InMemoryConventionBundleMasterTest {
     AssertJUnit.assertEquals(InMemoryConventionBundleMaster.IN_MEMORY_UNIQUE_SCHEME.getName(), uid3M.getScheme ());
     AssertJUnit.assertEquals(actact, conventions2.getDayCount());
     AssertJUnit.assertEquals(modified, conventions2.getBusinessDayConvention());
-    AssertJUnit.assertEquals(2, conventions2.getSettlementDays());
+    AssertJUnit.assertEquals(2, (int) conventions2.getSettlementDays());
 
     assertFalse(uidON.equals (uid3M));
